ci: update Docker tagging in GitHub Actions
This commit modifies the semantic-release workflow to adjust the Docker
image tagging strategy. Images will now be tagged with the full version,
the major version with `.x` (if the major version is not `0`), and
either the `latest` or `next` tag based on whether the version includes
a `-`. This aligns with our versioning policy and aids in Docker image
management.