ci: adjust semantic release in CI config
Modify the 'Print next version' and 'Python Semantic Release' steps in
the GitHub CI configuration to handle manual dispatch events
differently. For manual dispatch events on the 'release/*' branch, the
'force' option is not used. This change ensures that the semantic
release process behaves as expected for different types of workflow
events and branches.